propchain_traits::di

Trait ServiceRegistry

pub trait ServiceRegistry: ContractEnv {
    type registerServiceOutput: ImpliesReturn<Result<(), DependencyError>>;
    type unregisterServiceOutput: ImpliesReturn<Result<(), DependencyError>>;
    type resolveServiceOutput: ImpliesReturn<Result<AccountId, DependencyError>>;
    type isServiceRegisteredOutput: ImpliesReturn<bool>;

    // Required methods
    fn register_service(
        &mut self,
        key: ServiceKey,
        address: AccountId,
    ) -> Self::registerServiceOutput;
    fn unregister_service(
        &mut self,
        key: ServiceKey,
    ) -> Self::unregisterServiceOutput;
    fn resolve_service(&self, key: ServiceKey) -> Self::resolveServiceOutput;
    fn is_service_registered(
        &self,
        key: ServiceKey,
    ) -> Self::isServiceRegisteredOutput;
}

ink! trait for contracts that expose a service registry.

Implement this on your contract to provide a standard interface for registering, unregistering, and resolving service dependencies. Admin-gating of register_service / unregister_service is the responsibility of the implementing contract.

Required Associated Types

type registerServiceOutput: ImpliesReturn<Result<(), DependencyError>>

Output type of the respective trait message.

type unregisterServiceOutput: ImpliesReturn<Result<(), DependencyError>>

Output type of the respective trait message.

type resolveServiceOutput: ImpliesReturn<Result<AccountId, DependencyError>>

Output type of the respective trait message.

type isServiceRegisteredOutput: ImpliesReturn<bool>

Output type of the respective trait message.

Required Methods

fn register_service( &mut self, key: ServiceKey, address: AccountId, ) -> Self::registerServiceOutput

Register a service address for the given key.

Only callable by the contract admin.

fn unregister_service( &mut self, key: ServiceKey, ) -> Self::unregisterServiceOutput

Unregister a service.

Only callable by the contract admin.

fn resolve_service(&self, key: ServiceKey) -> Self::resolveServiceOutput

Resolve a service address by key.

Returns Err(DependencyError::ServiceNotRegistered) if not set.

fn is_service_registered( &self, key: ServiceKey, ) -> Self::isServiceRegisteredOutput

Returns true if the service is currently registered.